Added TOC page, welcome page, and cover page support to the Maker
Ensures we generate appropriate metadata for a TOC html page and a cover
art.  This has also been tested with the Kindle which relies and an
html toc page existing.

The cover can either be an image or HTML both should work.

Finally the first file in the manifest is set in the guide as the 'text'
type.  This is used by Kindle to determine where to start the book and
is considered the 'welcome' page.

The example has been updated to include the toc page and cover art